UserManagement.Web/Controllers/AccountController.cs
using Microsoft.AspNetCore.Identity;
using System.Threading.Tasks;
using UserManagement.Models;
using UserManagement.Models.Account;

namespace UserManagement.Controllers
{
    public class AccountController : Controller
    {
        private readonly SignInManager<User> _signInManager;
        private readonly UserManager<User> _userManager;

        public AccountController(SignInManager<User> signInManager, UserManager<User> userManager)
        {
            _signInManager = signInManager;
            _userManager = userManager;
        }

        [HttpGet("account/login/")]
        public IActionResult Login()
        {
            return View(new LoginViewModel { Email = string.Empty, Password = string.Empty});
        }

        [HttpPost]
        public async Task<IActionResult> Login(LoginViewModel model)
        {
            if (ModelState.IsValid)
            {
                var user = await _userManager.FindByEmailAsync(model.Email);
                if (user != null)
                {
                    if (user.UserName != null)
                    {
                        var result = await _signInManager.PasswordSignInAsync(user.UserName, model.Password, false, false);
                        if (result.Succeeded)
                        {
                            return RedirectToAction("Index", "Home");
                        }
                    }
                }
                ModelState.AddModelError("", "Invalid login attempt");
            }
            return View(model);
        }

        [HttpGet("account/logout/")]
        public async Task<IActionResult> Logout()
        {
            await _signInManager.SignOutAsync();
            return RedirectToAction("Login", "Account");
        }
    }
}
